Add files via upload
The most important feature while working on this dataset is that we deal with the Human Mind. Because from the process of thinking about an article, deciding to buy, arranging the money and waiting till the time of delivery is a wholesome process but if it doesn’t flow according to the promises and dates assigned it may affect the business for customer retention.